fix inactive sub card
